seperated bid function into bids.js
